refactor and bug fix
- move Intent class outside PlayerIntent
- exceptions in Strategy now throw its orginal exceptions
- some asDTO functions
- Game.Update now return GameDTO object

+++ kombat-backend/src/main/java/com/oop11/kombat_backend/Games/DTO/PlayerDTO.java
@@ -7,8 +7,8 @@ import lombok.extern.jackson.Jacksonized;
 public record PlayerDTO(
 	String name,
 	int team,
-	int budget,
-	int interestRatePercentage,
+	double budget,
+	double interestRatePercentage,
 	int spawnCount,
 	int minionCount)
 {
